% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbotUI.R
\name{chatbotUI}
\alias{chatbotUI}
\title{chatbotUI Shiny UI Module}
\usage{
chatbotUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
An user interface for your shiny watson assistant apps
}
\examples{
if(interactive()){
### APP ###
library(shiny)
library(shinydashboard)
ui <- dashboardPage(
  dashboardHeader(),
    dashboardSidebar(),
      dashboardBody(uiOutput("context"),
                      IMWatson::chatbotUI("watson"))
                      )
server <- function(input, output, session) {
rv <- reactiveValues(rv.chat = FALSE)
bot <- reactive({
callModule(IMWatson::chatbox, "watson", texlog)
})
output$context <- renderUI(bot()$magic)
}
shinyApp(ui, server)
 }
}
