% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_chat.R
\name{continue_chat}
\alias{continue_chat}
\title{Continue Chat with Watson}
\usage{
continue_chat(text, context, url, version, api_key, workspace)
}
\arguments{
\item{text}{Whatever you want to say to Watson}

\item{context}{context returned by your assistant}

\item{url}{optional argument that defaults to "https://gateway.watsonplatform.net/assistant/api/v1/workspaces"}

\item{version}{optional argument that defaults to "2018-09-20"}

\item{api_key}{api key to connect to your watson assistant}

\item{workspace}{Workspace api}
}
\value{
a list with the response and context produced by our assistant
}
\description{
Chat with Waston after initializing a connection
}
\details{
continue your conversation with your assistant
}
\examples{
if(interactive()){
 connection <- init(api_key, workspace)
 connection[["text"]]
 chat1 <- chat("Ignacio", connection, api_key, workspace)
 chat1[["text"]]
 chat2 <- chat("I would like flower suggestions", chat1, api_key, workspace)
 chat2[["text"]]
 }
}
