% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyChatbot.R
\name{ShinyChatbot}
\alias{ShinyChatbot}
\title{Shiny Chatbot}
\usage{
ShinyChatbot(api_key, workspace, title = "Chatbot powered by Watson")
}
\arguments{
\item{api_key}{The API key}

\item{workspace}{The workspace api for your chatbot}

\item{title}{The title for the chatbox}
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
if(interactive()){
library(IMWatson)
ShinyChatbot(api_key = your_api_key,
workspace = your_workspace)
 }
}
\seealso{
\code{\link[shiny]{runApp}}
}
