% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson_chat_factory.R
\docType{data}
\name{watson_chat_factory}
\alias{watson_chat_factory}
\title{watson chat factory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
watson_chat_factory
}
\description{
watson chat factory
}
\keyword{datasets}
